+++ src-gui/src/renderer/components/App.tsx
@@ -3,7 +3,7 @@ import { ThemeProvider } from "@material-ui/core/styles";
 import "@tauri-apps/plugin-shell";
 import { Route, MemoryRouter as Router, Routes } from "react-router-dom";
 import Navigation, { drawerWidth } from "./navigation/Navigation";
-import HelpPage from "./pages/help/HelpPage";
+import SettingsPage from "./pages/help/SettingsPage";
 import HistoryPage from "./pages/history/HistoryPage";
 import SwapPage from "./pages/swap/SwapPage";
 import WalletPage from "./pages/wallet/WalletPage";
@@ -14,6 +14,7 @@ import { themes } from "./theme";
 import { useEffect } from "react";
 import { setupBackgroundTasks } from "renderer/background";
 import "@fontsource/roboto";
+import FeedbackPage from "./pages/feedback/FeedbackPage";
 
 const useStyles = makeStyles((theme) => ({
   innerContent: {
@@ -54,7 +55,8 @@ function InnerContent() {
         <Route path="/swap" element={<SwapPage />} />
         <Route path="/history" element={<HistoryPage />} />
         <Route path="/wallet" element={<WalletPage />} />
-        <Route path="/help" element={<HelpPage />} />
+        <Route path="/settings" element={<SettingsPage />} />
+        <Route path="/feedback" element={<FeedbackPage />} />
         <Route path="/" element={<SwapPage />} />
       </Routes>
     </Box>