+++ src-gui/src/store/middleware/storeListener.ts
@@ -3,9 +3,10 @@ import { getAllSwapInfos, checkBitcoinBalance, updateAllNodeStatuses, fetchSelle
 import logger from "utils/logger";
 import { contextStatusEventReceived } from "store/features/rpcSlice";
 import { addNode, setFetchFiatPrices, setFiatCurrency } from "store/features/settingsSlice";
-import { updateRates } from "renderer/api";
+import { fetchFeedbackMessagesViaHttp, updateRates } from "renderer/api";
 import { store } from "renderer/store/storeRenderer";
 import { swapProgressEventReceived } from "store/features/swapSlice";
+import { addFeedbackId, setConversation } from "store/features/conversationsSlice";
 
 export function createMainListeners() {
   const listener = createListenerMiddleware();
@@ -75,6 +76,16 @@ export function createMainListeners() {
     actionCreator: addNode,
     effect: async (_) => {
       await updateAllNodeStatuses();
+    },
+  });
+
+  // Listener for when a feedback id is added
+  listener.startListening({
+    actionCreator: addFeedbackId,
+    effect: async (action) => {
+      // Whenever a new feedback id is added, fetch the messages and store them in the Redux store
+      const messages = await fetchFeedbackMessagesViaHttp(action.payload);
+      store.dispatch(setConversation({ feedbackId: action.payload, messages }));
     },
   });